feat: convert to typescript
Converted the components from JavaScript to TypeScript.
