+++ package.json
@@ -9,6 +9,12 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.7.1",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.33",
+    "@types/react": "^17.0.3",
+    "@types/react-dom": "^17.0.2",
+    "@types/react-router-dom": "^5.1.7",
+    "@types/yup": "^0.29.11",
     "axios": "^0.21.1",
     "formik": "^2.2.6",
     "node-sass": "^5.0.0",
@@ -16,6 +22,7 @@
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.2",
+    "typescript": "^4.2.3",
     "web-vitals": "^1.1.0",
     "yup": "^0.32.9"
   },
