+++ src/App.tsx
@@ -1,7 +1,7 @@
 import "./App.scss";
 import Index from "./container/Index";
 import Contact from "./container/Contact";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Route } from "react-router-dom";
 import ScrollToTop from "./components/ScrollToTop";
 
 function App() {
