+++ src/interfaces/interface.ts
@@ -1,0 +1,9 @@
+export interface FormValues {
+    name: string,
+    email: string,
+    message: string, 
+}
+
+export interface ContactFormProps {
+    handleSubmit: (arg: FormValues) => void;
+  }
